backend/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
flask-cors==3.0.10
python-dotenv==0.21.0
openai>=1.0.0
PyPDF2==3.0.1
python-pptx==0.6.21